Fixed Pokemon best IV's where the top two statproducts are equal, algorythm now uses the Iv's with the best IV spread. So 15/15/15 and 15/15/14 produce the same stat product, we'll use the 15/15/15/. In the future I'd like to display these IV;'s to show people that 15/15/14 and 15/15/15 have the exact same stats so therefore either of them would be considered the same because the stamina stat is rounded so even a 14 stamina IV may end up being the same value as 15 stamina IV once it's rounded. That's how we get these pokemon that are 100% in every way, stats cp level etc, except their stamina stat is 14 instead of 15. If you have these pokemone then you have a hundo 100%

+++ GoBattleLeagueTeamBuilderProject/GoBattleLeagueTeamBuilder/Models/IVPerformance.cs
@@ -8,24 +8,24 @@ namespace GoBattleLeagueTeamBuilder.Models
     public class IVPerformance
     {
         public IV iVS { get; set; }
-        public float level { get; set; }
+        public double level { get; set; }
         public stat stats { get; set; }
         public int cP { get; set; }
-        public float statProduct { get; set; }
+        public double statProduct { get; set; }
 
     }
 
     public class IV
     {
-        public float atkIV { get; set; }
-        public float defIV { get; set; }
-        public float staIV { get; set; }
+        public double atkIV { get; set; }
+        public double defIV { get; set; }
+        public double staIV { get; set; }
     }
 
     public class stat
     {
-        public float atkStat { get; set; }
-        public float defStat { get; set; }
-        public float staStat { get; set; }
+        public double atkStat { get; set; }
+        public double defStat { get; set; }
+        public double staStat { get; set; }
     }
 }
