+++ GoBattleLeagueTeamBuilderProject/GoBattleLeagueTeamBuilder/Models/Interfaces/IPVP_IVsAPIRepository.cs
@@ -7,10 +7,10 @@ namespace GoBattleLeagueTeamBuilder.Models.Interfaces
 {
     public interface IPVP_IVsAPIRepository
     {
-        void PrintBestIV(int league, int baseAtk, int baseDef, int baseSta, float xLStatus);
+        void PrintBestIV(int league, int baseAtk, int baseDef, int baseSta, double xLStatus);
        
 
-        IVPerformance GetBestIV(int league, int baseAtk, int baseDef, int baseSta, float xLStatus);
+        IVPerformance GetBestIV(int league, int baseAtk, int baseDef, int baseSta, double xLStatus);
        
     }
 }
